backend/requirements.txt
Flask==2.0.1
Flask-CORS==3.0.10
Flask-SQLAlchemy==2.5.1
PyJWT==2.1.0
passlib==1.7.4
python-dotenv==0.19.0